requirements.txt
Flask>=2.0
python-nmap>=0.7.1
flask-cors>=3.0
